+++ lib/addressfinder.rb
@@ -9,6 +9,8 @@ require 'addressfinder/address_info'
 require 'addressfinder/address_search'
 require 'addressfinder/address_autocomplete'
 require 'addressfinder/bulk'
+require 'addressfinder/v1/email/verification'
+require 'addressfinder/v1/phone/verification'
 require 'addressfinder/errors'
 require 'addressfinder/util'
 require 'addressfinder/http'
@@ -59,6 +61,14 @@ module AddressFinder
 
     def address_info(args={})
       AddressFinder::AddressInfo.new(params: args, http: AddressFinder::HTTP.new(configuration)).perform.result
+    end
+
+    def email_verification(args={})
+      AddressFinder::V1::Email::Verification.new(**args.merge(http: AddressFinder::HTTP.new(configuration))).perform.result
+    end
+
+    def phone_verification(args={})
+      AddressFinder::V1::Phone::Verification.new(**args.merge(http: AddressFinder::HTTP.new(configuration))).perform.result
     end
 
     def bulk(&block)